Refactor IP patterns validation
